+++ .gitignore
@@ -3,4 +3,5 @@ node_modules/
 out/
 dist/
 .vscode-test
+.vscode-test-web
 *.vsix
