+++ package.json
@@ -45,10 +45,11 @@
     "pretest": "npm run test-compile",
     "test": "node ./out/test/runTest.js",
     "lint": "node ./node_modules/tslint/bin/tslint --project tsconfig.json -t verbose",
-    "lint-fix": "tslint --project tsconfig.json -t verbose --fix"
+    "lint-fix": "tslint --project tsconfig.json -t verbose --fix",
+    "open-in-browser": "vscode-test-web --extensionDevelopmentPath=. ."
   },
   "dependencies": {
-    "axios": "^1.7.4",
+    "axios": "^1.8.1",
     "node-cmd": "^5.0.0"
   },
   "devDependencies": {
@@ -58,6 +59,7 @@
     "@types/node": "^20.14.0",
     "@types/vscode": "^1.93.0",
     "@vscode/test-electron": "^2.3.8",
+    "@vscode/test-web": "^0.0.67",
     "glob": "^7.1.6",
     "mocha": "^11.0.1",
     "rimraf": "^5.0.5",
@@ -68,6 +70,7 @@
     "webpack-cli": "^5.1.4"
   },
   "main": "./dist/extension",
+  "browser": "./dist/web-extension",
   "activationEvents": [],
   "contributes": {
     "authentication": [
